chore: move more components into `case-tile-common`
